requirements.txt
pandas==2.0.3
python_docx==1.1.0
streamlit==1.27.2
